Separate test cases into multiple files.
